test(cgroups): split tests in several files
